// src/interfaces/championship-table.ts
export interface ChampionshipTable {
  campeonato: string;
  posicao: number;
  equipe: string;
  pontos: number;
  maximoPontosPossivel: number;
  jogos: number;
  vitorias: number;
  empates: number;
  derrotas: number;
  golsPro: number;
  golsContra: number;
  saldoGols: number;
  aproveitamento: number;
  vitoriasMandante: number;
  derrotasMandante: number;
  empatesMandante: number;
  vitoriasVisitante: number;
  empatesVisitante: number;
  derrotasVisitante: number;
  SiglaTime: string;
  URLlogo: string;
  apelidoEquipe: string;
}